API - Refactored exceptions in a single file
