--- app/libs/exceptions/argumentException.py
@@ -1,5 +1,0 @@
-class ArgumentException(Exception):
-
-	def __init__(self, message):
-		super().__init__(message)
-		self.message = message
